Fix //src/test/shell/integration:progress_reporting_test
The test case test_failed_actions_with_keep_going was never correct
and only worked by accident: it relied on bazel starting at least
two actions, if for there are at least two where all the dependencies
are already met; however bazel is clever and analyzes the resources
of the system it is running in, and if they are small will only run one
genrule at a time. Fix this by telling it to run 2 actions and lying
about the available resources of the local system.

Also increase the test size and reenable the test for remote execution.
Fixes #3995.

Change-Id: Ic84de17a98a2e74c24901d4efb38d7631de4690a
PiperOrigin-RevId: 176511026