Return SpawnResults using a List instead of a Set.
Currently we don't care about the list order of SpawnResults. However, we may care about the list order later. Also, if the equals() method for SpawnResults is ever changed then it may be problematic to return SpawnResults in a Set.

Aside: ActionResults use SpawnResults to calculate cumulative execution times for Actions, and may provide other metrics in future.

RELNOTES: None.
PiperOrigin-RevId: 176579460