Automated rollback of commit 0ebb3e54fc890946ae6b3d059ecbd50e4b5ec840.
*** Reason for rollback ***

Cause rare case of test hanging.

To cherry-pick for #3772

*** Original change description ***

Attempting to fix an occasionally missing stdout from test.xml.

In hello-world_test, when executed inside a docker container, for about 2% of the runs the test.xml has an empty CDATA, instead of the expected "Hello, world!". I'm not sure still what exactly was the bug, but in any case this change simplifies the test execution code line, so if this doesn't fix it, at least further debugging will be easier.

I ran the test in a loop 1360 times, and the error did not reproduce once, which hints to...

***

PiperOrigin-RevId: 176662654