refactor: add LocationExpander factories
Stop using LocationExpander.Options in
LocationExpander constructors, because the Options
semantics are confusing.

I also need the refactoring in order to extend the
expansion semantics: to support expanding to
absolute paths on Windows, where $(location)
should not expand to the (non-existent) runfiles
path, but to the absolute path the runfiles
symlink would point to.

See https://github.com/bazelbuild/bazel/issues/4171

Change-Id: Ie4d47ec3807bc3c6e39156efa1746b666f69f99c
PiperOrigin-RevId: 177147372