Automated rollback of commit b64119807b014d9f3b99fb8a02e22daf1a8299b6.
*** Reason for rollback ***

Tickles the shell integration test version of b/35042288

*** Original change description ***

Change BlacklistedPackagesPrefixesFunction to take a pair of a hardcoded set of directories and a file path containing more directories to blacklist. The current usage of PrecomputedValue#BLACKLISTED_PACKAGE_PREFIXES_FILE is overly general and is only meaningfully used by unit tests; in practice, the blacklist file path can never change over the lifetime of the Bazel server. Perform a minor simplifying refactor as a result of this.

RELNOTES: None.
PiperOrigin-RevId: 177176068