Handle spaces in absolute paths to tests.
This allows for output_base paths that contain spaces, fixing #3897

The test path is set by grepping the MANIFEST, in which each line is a
space-separated 2-tuple of relative path and absolute path to the test.
Replace the awk command that extracts the second item in the
space-separated list with a sed command that extracts everything after the
first space. Also, quote the manifest path arg to the grep command.

There will still be problems if there are spaces in the relative path,
below the Bazel project root.

Test: Built and tested a test with an output_base containing a space.

Change-Id: I5dba285bb50c27eedace17a5e20efac7392a9a6e

Closes #3899.

Change-Id: I5dba285bb50c27eedace17a5e20efac7392a9a6e
PiperOrigin-RevId: 177283695