Expressly fetch remote and ref when using git_repository
When I clone from a git repo that doesn't have the default branch the same as the branch that I want ; it doesn't seem to fetch the appropriate branch. I am using a name rather than explicit sha.

What happens is:
```
+ git reset --hard build-with-bazel
fatal: ambiguous argument 'build-with-bazel': unknown revision or path not in the working tree.
```

because it hasn't fetched that ref.

Closes #4039.

PiperOrigin-RevId: 177298578