Buffer blaze query
It was previously sending each label individually over gRPC, where each call
has a lot of overhead.

This makes queries with a large amount of output _a lot_ faster. For an example
query where all packages are already loaded, I observe a difference of ~3.5s
before this change to ~1.6s after this change.

PiperOrigin-RevId: 177426957