remote: don't hide non-test failures behind test failures. Fixes #4082
Bazel should display the root cause of a test failure to the user. For
example, if a test could not be executed on a remote executor due to
there being no network connection, then it shouldn't display the test
as failed but tell the user about the network error.

RELNOTES:
PiperOrigin-RevId: 177439578