py_binary can now be used as data with custom python.
The stub template now looks for the python binary relative to the
modules if a relative path was provided.  This correctly finds it
inside the runfiles folder both when the py_binary is the output,
and when the py_binary is called by another binary
(ie is data for it).

We also now add the binary and dependencies to the runfiles when it
is used as data so python is accesible.

Change-Id: I3bf6ff17265e72d964614ad66af22933c89f853d
PiperOrigin-RevId: 177803641