Expose ObjcProtoProvider as a Skylark object.
This does not provide any meaningful API to interact with ObjcProtoProvider from a skylark context -- it simply allows ObjcProtoProvider to be passed between skylark API calls as an opaque object.

This helps facilitate exposure of an Apple Linking API to Skylark which will take place in future changes.

RELNOTES: None.
PiperOrigin-RevId: 177844029