Remove static import of inner class InferredType's fields
A few of the InferredType's fields were being static imported, but not all of them, and the use of the statically imported fields was inconsistent.  Statically importing the inner class causes strange errors when building desugar in the Android platform build with OpenJDK8, but not OpenJDK9:
external/desugar/java/com/google/devtools/build/android/desugar/BytecodeTypeInference.java:1015: error: cannot find symbol
  @AutoValue
   ^
  symbol:   class AutoValue
  location: class BytecodeTypeInference
1 error

Remove the static imports to make the build work and to make the usage consistent.

RELNOTES:None.
PiperOrigin-RevId: 177875501