Add tests for circular target dependencies resolved by configuration.
It's possible for a change in the configuration to result in multiple
visits to a particular target, but as long as each visit has a unique
configuration (relative to other visits to that target) and the cycle
is eventually broken (i.e., one of the targets in the cycle does not
depend on the next target in the cycle if the configuration satisfies
some property), this is perfectly buildable.

Or, in other words, it's okay to have a cycle in the target graph as
long as it is not also a cycle in the _configured_ target graph.

This test is to confirm this property.

RELNOTES: None.
PiperOrigin-RevId: 178303838