Make config_feature_flag's default_value optional.
With this change, if the default_value is not set, but the flag
has been given a value by a configuration transition, that value
is used and no error is produced.

If the default_value is not set and the flag has not been given
a value by a configuration transition, the rule produces an
error (as it did before, but now the error is different).

config_feature_flags with default_values behave as they did before.
(use the default value if no configured value is set, otherwise use
the configured value)

Also transition the Optional used in feature flags from Guava to
Java 8.

RELNOTES: config_feature_flag's default_value is optional. It is
only an error to have a config_feature_flag with no default_value
if that config_feature_flag has not been set in the configuration
it is being evaluated in.
PiperOrigin-RevId: 178418907