Fixed documentation navigation bar/content layout awkwardness.
Fixes https://github.com/bazelbuild/bazel/issues/4262

A specific browser view width range causes the navigation bar to be stacked on top of the documentation content, therefore requiring a bit of scrolling to get to it.

This fix changes col-lg-* to col-md-*, which reduces the view width threshold at which the columns  are restyled using `@media queries`.

Tested locally on Chrome & Safari. This is also testable by going to the production site and changing the CSS classes in the inspector.

Closes #4263.

PiperOrigin-RevId: 178605473