Automated rollback of commit 9321316b34767b06c3071b2cf2a4de189874fcce.
*** Reason for rollback ***

Design change, 2 boolean flags instead of 1 enum flag

*** Original change description ***

Add --incremental_state_retention_strategy

This option is intended to replace some of the uses of --batch. It lets users specify that builds should not be incremental, and how eagerly to discard the state that is kept around for incrementality. Note that for both values discard_eargerly and keep_for_life_of_build, the build graph is kept around until the next build. This may change.

Will add tests for keep_for_life_of_build in a later change, for now it will warn that that feature is experimen...

***

ROLLBACK_OF=178661777
RELNOTES: None.
PiperOrigin-RevId: 178681472