Add a new file path type, LocalPath.
This path type is a local file path as a wrapper around a string. It works much the same as java.io.File, but without its file operations.

For the most part, FilePath shouldn't add much overhead compared to using plain strings. Strings do get normalised on the way in, but no extra objects are allocated unless the path actually needs normalisation.

PiperOrigin-RevId: 178798497