Remove source specific build variables from C++ API.
The only use case was for Objective-C rules so that different set of sources specified in the same rule could be compiled with either ARC or no ARC. To replace source specific build variables, we call into CcLibraryHelper twice for each set of sources.

This has led to separating the building of compilation providers and outputs from those related to linking. In the case of Objective-C, the compilation outputs are merged and then passed to a single invocation of linking in CcLibraryHelper.

In a later CL, the distinction between compilation and linking will be refined by separating them in different classes.

RELNOTES:none
PiperOrigin-RevId: 179167102