third_party: update google http client
Update the google http client to a custom built binary. It's the sources
from version 1.22.0 with one cherry-pick [1]. The cherry-pick makes
google http client compatible with the apache http client version we are
using.

The next release of google http client will contain this commit and from
there on we can just stick to their release version again.

[1]
https://github.com/google/google-http-java-client/commit/4b5f59ed042d3b55b3148a1b01f1d5ee0ad7460c

Change-Id: If6273be78b716dd417a60096599c036cc8101efb