bazel_bootstrap_distfile_test: run a more simple setup
The bootstrap-distfile test aims at verifying the completeness of
the distribution archive. The functionality of workers is a different
issue, so separate those concerns and do the test bootstrap in a
simple setup, calling javac standalone (see #1250). Also, reduce
the output to make the test log more readable.

Improves on #4306.

Change-Id: I839523628cd35f2deefc1a4dfdf0cf58a4d1f51a
PiperOrigin-RevId: 179186615