look for the prelude in each repository (fixes #1673)
Instead of every package using the root repository's prelude, it is resolved in each repository for the packages in that repository.

There weren't any tests using the prelude at all, so these are better than nothing.  I wrote them as bash integration tests instead of Java tests because I couldn't get the prelude to work at all with `BuildViewTestCase` -- I'd be happy to share the busted test I wrote if someone would like to help debug it.

Closes #3991.

PiperOrigin-RevId: 179557018