Use more robust check for bitcode inputs to link
The current checks used in several places for whether an LTO indexing
step is needed only looked for direct LTO bitcode inputs. Instead we should
use the existing hasLtoBitcodeInputs() method to look both at the direct
inputs as well as inputs on all LibraryToLink. This was already used in
similar checking on the CppBinary. With this change, all places that
test for needing an LTO indexing step use the correct interface.

RELNOTES: None
PiperOrigin-RevId: 179812972