Terminate --javacopt flags with `--`
JDK 9 adds javac flags that start with `--`, so we can't assume that `--` flags
belong to JavaBuilder or turbine. Instead, we will use `--` by itself as a delimiter
to parse e.g. `--javacopts --release 8 -- sources Hello.java`.

This change cannot be released until versions of turbine and JavaBuilder are released
that support the trailing `--`. After it is included in a Blaze release the other tools
will be updated again to require the trailing `--`.

PiperOrigin-RevId: 179847276