Move config transition definitions out of lib.packages.Attribute.
This cleans up a legacy API from Bazel's "static configuration" days,
when transition definitions involved deep and complex logic in the
analysis code. That was too much cruft to embed into lib.packages.
But the modern definitions are way simpler and thus easier to embed.

This change only *copies* the existing definitions. Because a lot of
references will have to be updated, we'll migrate the move over a few
changes.

PiperOrigin-RevId: 179859293