Intern the BuildConfigurationValue.Key instances that we create. Currently we don't create too many of them, but it'll be nice to have this interning for future work in which there are many more of them.
When comparing BuildConfiguration instances, compare their full BuildOptions fields, which can compare quickly due to fingerprint caching, as opposed to their raw options, which is slow.

Also intern the map of Fragments that we create as part of a BuildConfiguration. There aren't too many of them either, but it means that equal Fragment sets can be compared using reference equality downstream.

PiperOrigin-RevId: 180289334