Use -no-canonical-prefixes when -fno-canonical-system-headers is not available
Bazel's autoconf script adds -fno-canonical-system-headers to gcc's crosstool
in order to get output in '.d' files that can be properly verified by bazel.
To workaround the same issue in clang we have to use a different flag:
'-no-canonical-prefixes'.

The same issue arises with clang if it resides in the configured repository
(e.g., when it is downloaded via 'repository_ctx.download').

PiperOrigin-RevId: 180552155