Rewrite //src/test/py/bazel:action_temp_test
Merge the test methods into one test method, in
order to reuse the workspace and the Bazel server.

Add an input file to the mock build rules and
update it before each build, to ensure that Bazel
reruns the actions and uses the strategy-specific
temp directory.

This change reduces the test's runtime on Linux
from 20.2s to 14.7s (average of 5 runs each).

Change-Id: Ia2a8d477d13271af59cdf55cdc49f137460a1b96
PiperOrigin-RevId: 180796797