Merge the runfiles of @androidsdk//:aapt_binary into android_instrumentation_test's runfiles, and pass the execpath of aapt_binary to the entry point using the --aapt flag.
This allows android_instrumentation_test to use the dynamically linked AAPT from the user-provided SDK.

The reason for not using %aapt% template expansion in the stub template is because @androidsdk//:aapt_binary is a shell script with hardcoded paths to the actual dependencies in its runfiles directory, but since this binary is called within an android_instrumentation_test's runfiles directory, the nested aapt runfiles directory doesn't exist because runfiles are not nested, but merged.

We can make this work by referring to the actual external/androidsdk/aapt_binary using a `../../` directory traversal, and we can also assume that the aapt_binary is there because android_instrumentation_test merged aapt's runfiles from the SDK.

RELNOTES: None.
PiperOrigin-RevId: 180888739