Fix bug in unix_jni.cc that causes Path#exist to spuriously return true.
We use errno to signal an error in the stat call that gets passed back to Java land. However, between the time we make the failed stat syscall and the time that we read the final value of errno we will very rarely make other syscalls in between, which will stomp the value of errno back to 0. This will get interpreted as "no error, the file exists" by the VFS.

This bug has existed since 2009. Only a perturbation of the sequence of syscalls we make during startup has since caused the bug to surface.

PiperOrigin-RevId: 181358035