Add support for strip_prefix to {new_}git_repository skylark rules
Continuation of https://github.com/bazelbuild/bazel/pull/4356

An approach at supporting strip_prefix with the git skylark rules.

This approach unfortunately uses symlinks since you cannot clone a subset of a git repository. It creates a tmp directory which is the 'real' clone and then provides a link in place of the expected location of the repository to the path of the required prefix. Behaviour is only changed if a strip_prefix is provided.

Closes #4368.

PiperOrigin-RevId: 181438640