Remove package loading from JvmConfigurationLoader .
This should not cause a user-visible change: --{host_,}javabase was already
required to point to a java_runtime_suite / java_runtime rule, the only change is that
we now enforce it differently.

If anything, this makes Bazel more lenient in that it now accepts Skylark rules
that provide JavaRuntimeInfo, but I doubt anyone does that, since it does not
work at HEAD.

RELNOTES: None.
PiperOrigin-RevId: 181444563