Use an annotation preprocessor to validate SkylarkConfigurationField.
This is a rollforward -- the previous attempt was rolled back due to incorrectly adding test sources.
In addition, the newly checked in tests do not work on Windows, and are thus disabled on the Windows platform. I suspect this is due to a bug in the compile-testing library.

RELNOTES: None.
PiperOrigin-RevId: 181482589